Add requirement on Ruby version 2.5+
A recent commit using StandardRB to reformat sources dropped the "begin"
in "rescue" and "ensure" blocks, but using those keywords in blocks
without a "begin" is only available in Ruby 2.5.0 or later.

https://www.ruby-lang.org/en/news/2017/12/25/ruby-2-5-0-released/

Add a requirement on version of Ruby to be >= 2.5.0 to the gemspec.

I suggest bumping up the version (to 1.0.1 or later) to release a new
gem including this requirement. Retiring version 1.0.0 which doesn't
include the Ruby version requirement would be a good idea, that way
gem/bundler will be able to properly resolve dependencies to pick the
latest version that works with the running Ruby.

Tested:

  $ ruby -v
  ruby 2.4.9p362 (2019-10-02 revision 67824) [x86_64-linux]
  $ gem install ./climate_control-1.0.1.gem
  ERROR:  Error installing ./climate_control-1.0.1.gem:
          climate_control requires Ruby version >= 2.5.0.

Failure while running rake/rspec on Ruby 2.4:

  An error occurred while loading ./spec/acceptance/climate_control_spec.rb.
  Failure/Error: require "climate_control/environment"

  SyntaxError:
    ./lib/climate_control/environment.rb:22: syntax error, unexpected keyword_ensure, expecting keyword_end
	  ensure
		^
    ./lib/climate_control/environment.rb:33: syntax error, unexpected keyword_end, expecting end-of-input
  # ./lib/climate_control.rb:1:in `<top (required)>'
  # ./spec/spec_helper.rb:13:in `<top (required)>'
  # ./spec/acceptance/climate_control_spec.rb:1:in `<top (required)>'
  No examples found.