Refactor: use Montior instead of Mutex, prefer locals over ivars
`Monitor` ships with the standard library and is reentrant,
simplifying the code in lib/climate_control/environment.rb quite a bit.
https://rdoc.info/stdlib/monitor/Monitor

I also looked at the code in lib/climate_control/modifier.rb
and refactored it towards local variables because I found it
difficult to keep track of all the instance variables.

The `to_hash` is a little sketchy because it works with `ENV`,
but not with normal hashes:

```bash
ruby -e '
  def show_to_hash(h)
    p [h.object_id, h.to_hash.object_id]
  end

  show_to_hash({})
  show_to_hash(ENV)
'
[60, 60]
[80, 100]
```

But I was having issues using `dup`, so I ultimately left it alone.