Remove ClimateControl::Environment
After introducing `monitor` from the stdlib, it was only useful as a
reference to the environment hash and the monitor, which didn't seem
necessary anymore.

I also chose a constant over a class variable, because class variables
can behave in very unexpected ways.

I also changed `Modifier#process` to `Modifier#call` because that's the
convention in Ruby.