Allow ClimateControl.modify to be called without environment variables
### Why was this change necessary?

When wanting to wrap some block inside ClimateControl for thread safety
for instance it's useful to be able to just call:

`ClimateControl.modify { }` or `with_modified_env {}`

See #38 for an example.

### How does it address the problem?

This changes allows `.modify` to be called without an environment
variables hash (defaults to empty hash `{}`).

### Are there any side effects?

Some applications could theoretically be relying on the fact that
calling .modify without an environment variables hash would raise an
error. I think it's very unlikely.