Add documentation about thread-safety
### Why was this change necessary?

It was unclear what was the thread safety guarantees of climate_control.
Thus this could lead to failing tests in the gem's user's test suite.

### How does it address the problem?

This documents the status of thread safety and how to make one's code
thread safe.

### Are there any side effects?

We are committing to a particular interface about thread safety. I think
it's a reasonable interface though.