ENV was not restored if an error was thrown when assigning ENV
### Why was this change necessary?

Due to the refactor (#41) a case was not handed correctly:

```ruby
ENV.values_at 'A', 'B'  # => ["pre a", "pre b"]
ClimateControl.modify(A: 'overridden a', B: :not_a_string) {} rescue nil
ENV.values_at 'A', 'B'  # => ["overridden a", "pre b"]
```

### How does it address the problem?

This ensures we restore the ENV even when the assignment of ENV values
failed.

### Are there any side effects?

This restores a previous behavior which was the intended and correct
one.

https://github.com/thoughtbot/climate_control/pull/41#discussion_r883983553

Co-authored-by: Josh Cheek <josh.cheek@gmail.com>